alx_travel_app/requirement.txt
django==4.2.11
djangorestframework==3.15.0
celery==5.3.6 
drf-yasg==1.21.7
django-cors-headers==4.2.0
mysqlclient==2.2.0
django-environ==0.11.2
amqp==5.2.0